+++ Tests/MapboxSearchUITests/Tests/FavoritesUITestCase.swift
@@ -1,7 +1,6 @@
 import XCTest
 
-// Rename to UITestCase
-class FavoritesIntegrationTestCase: MockSearchBoxUITestCase {
+class FavoritesUITestCase: MockSearchBoxUITestCase {
     func testAddRemoveFavorite() throws {
         try server.setResponse(.suggestMinsk)
         try server.setResponse(.retrieveMinsk)
@@ -233,7 +232,7 @@ class FavoritesIntegrationTestCase: MockSearchBoxUITestCase {
     }
 }
 
-extension FavoritesIntegrationTestCase {
+extension FavoritesUITestCase {
     func removeDefaultFavorite(element: XCUIElement) {
         element.buttons["moreButton"].tap()
         let removeLocation = app.buttons["Remove location"]