+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 <!-- Add changes for active work here -->
 
+- [Tests] Reorganize tests based on API type
+
 - [Privacy] Add Search history collected data for the purpose of product personalization (used for displaying the search history)
 
 - [Discover, Category] Rename Discover to Category and update tests.