+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 - [Tests] Change MockResponse into a protocol, create separate enums conforming to MockResponse for each API type (geocoding, sbs, autofill), add MockResponse as generic to each test base class and MockWebServer.
 
+**MapboxCoreSearch**: v2.0.0-alpha.14
+
 ## 2.0.0-rc.2
 
 - [Discover] Add support for country, proximity, and origin parameters in Discover.Options search parameters. This fixes an issue when using search-along-route to query category results.